requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.118
botocore==1.34.118
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-rest-knox==4.2.0
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
facebook-sdk==3.1.0
future==1.0.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.7
oauthlib==3.2.2
packaging==24.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
paystackapi==2.1.1
python-twitter==3.5
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
